// src/pages/modules/LPPM/services/data/types.tsx
export interface IServices {
  id_layanan: string
  id_satuan_organisasi: string
  nama_layanan: string
  url: string
  posisi_header: boolean
  posisi_bawah_landing: boolean
  posisi_footer: boolean
  urutan: number
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
}